refactor: move scene related util classes from util to new package 'scene.util'
