+++ src/main/java/io/github/somesourcecode/someguiapi/scene/layout/Region.java
@@ -23,7 +23,7 @@
 
 package io.github.somesourcecode.someguiapi.scene.layout;
 
-import io.github.somesourcecode.someguiapi.util.Insets;
+import io.github.somesourcecode.someguiapi.scene.util.Insets;
 import io.github.somesourcecode.someguiapi.scene.Parent;
 
 /**
